make generated code simpler and more robust to collisions

+++ tl-types/build.rs
@@ -16,7 +16,7 @@ type User id:int64 verified:bool rating:float = User
 type UserEmpty id:int64 = User
 
 error InvalidUserId user_id:int64
-error TooLongText text:string max_length:int64
+error TooLongText text:string max_length:int32
 
 func get_users user_ids:[int64] = [User]
 func send_message user_id:int64 text:string? photos:[bytes] = Message")
